Fix license metadata
The galaxy upload rejects GPLv3 and requires a valid SPDX license
tag. The file headers list:

  GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)

which is a way of saying "3 or later" - so use that SPDX tag.

Also - the license headers refer to the file being COPYING, so
rename the file and update it in the galaxy metadata to match.

Change-Id: I6c82b49d7ce5ea8c9cde939d5d37ff54b969a598