Cleanup functional tests
Use the openstacksdk base functional test instead of going all
the way back to devstack for our base. Also - defining a job
starting with openstacksdk in this repo is not ok - but that
probably just came from the original copy-pasta coming from the
sdk repo.

Use tox-siblings more and stop doing tox requirements of ../
things. The system is there and we can use it.

Add jobs to test against released openstacksdk and ansible, as well
as a non-voting to test against ansible devel branch.

Add openstacksdk to the test-requirements because it's super
required.

Move zuul.yaml to .zuul.yaml. This isn't a zuul jobs repo so
we should just make it a hidden file rather than a non-hidden.

Change-Id: I658bd1728629cd358bc17a7f00d282d04530d8fe