Revert openstack.cloud.server parameter user_data to userdata
openstacksdk's create_server() [1] takes cloud-init userdata as a
plaintext string (or other data types [2]) via argument 'userdata',
not 'user_data' [3]. This is 'userdata' argument of create_server()
is different from the 'user_data' value which is returned [4]. The
latter is base64 decoded, while the former is e.g. a plaintext string.
As a (tiny) visual indicator for this difference, this patch reverts
the module parameter 'user_data' to 'userdata' which was used before
this module got refactored [5].

[1] https://opendev.org/openstack/openstacksdk/src/commit/57fbb72e32351f795f5888306096e8d3d35a64ef/openstack/cloud/_compute.py#L678
[2] https://opendev.org/openstack/openstacksdk/src/commit/57fbb72e32351f795f5888306096e8d3d35a64ef/openstack/cloud/_compute.py#L1757
[3] https://opendev.org/openstack/openstacksdk/src/commit/57fbb72e32351f795f5888306096e8d3d35a64ef/openstack/cloud/_compute.py#L796
[4] https://opendev.org/openstack/openstacksdk/src/commit/57fbb72e32351f795f5888306096e8d3d35a64ef/openstack/compute/v2/server.py#L223
[5] https://opendev.org/openstack/ansible-collections-openstack/commit/ac401bb354c6a1d388848a3ebd18537beba080cc

Change-Id: I1f3b9314e4d82674cd2ae45f6209de5611e2b6cb