Remove MAC address requirement for node registration
Allow the creation of baremetal nodes without specifying any NICs.
In this case switch the target provision state from available to
enroll.

We're proposing this patch to support enrolling nodes as a first
step before running an out-of-band Redfish inspection to get the
NICs.

No change is made to existing workflows.

Change-Id: I9b1636d778e7477191bfbe6094bd6bc70a4d384e
Signed-off-by: Hollie Hutchinson <hollie@stackhpc.com>
Signed-off-by: Doug Szumski <doug@stackhpc.com>